+++ src/backend/components/card/Instant.h
@@ -9,7 +9,7 @@
 
 class Instant : public Card {
 public : 
-    Instant(uint8_t id, const std::string &name, std::map<Color, int> cost);
+    Instant(uint16_t id, const std::string &name, std::map<Color, int> cost);
     ~Instant();
 
     Instant(Instant const&) = delete;
