+++ src/backend/components/card/Land.h
@@ -9,7 +9,7 @@
 
 class Land : public Card {
 public : 
-    Land(uint8_t id, const std::string &name, Color color);
+    Land(uint16_t id, const std::string &name, Color color);
     ~Land();
 
     Land(Land const&) = delete;
